Enable MSVC to work.
Tested on MSVC 2017. This patches makes it viable to build standalone
and debuggable .dlls of vkd3d.

This commit fixes various compatibility issues with MSVC 2017.

The only non-trivial change is the addition of a trivial pthread
wrapper. This requires Windows Vista and up due to condition variable
support being non-existent in XP.