build: Rename exposed APIs to reflect vkd3d-proton and update to 2.0.
Rename so objects we build so we don't conflict with vkd3d and don't
accidentially attempt to be built against Wine natively (it won't work).

Not quite ready for a 2.0 release yet, but bump the version to reflect
the intent. This creates a new timeline, completely separate from vkd3d.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>