vkd3d: Rewrite how buffers are stored in desc heap metadata.
Rather than storing VkBuffer / offset / size, we now store:

- VA
- Range (32-bit)
- Format

This accomplishes several things:

- Prepares for a world where VkBufferView is not created in the heap itself (i.e. no offset buffer).
- The metadata struct can now be shrunk to a neat 16 bytes instead of 32 before the refactor began.
- Allow UAV clears to operate without a dependency on offset buffers.

Hoisted CBV table descriptors (barely used, if ever) need to deref the VA late, but it's okay.
Hopefully we get the option to push descriptors with VA later, but it's not critically important now.

ClearUAV paths need to be rewritten to account for this.
Rather than reusing VkBufferViews and reading the offset buffer,
we can just create the buffer view on-demand instead.
A VA deref is fortunately not needed since we get the ID3D12Resource inline.
We already had to create fallback views for fallback clear paths anyways.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>