vkd3d: Add vkd3d_native_sync_handle.
Abstracts OS differences for HANDLE synchronization.

Currently, there is a lot of cruft that was inherited from vkd3d
where event HANDLEs take callbacks.
When trying to add semaphore support for frame latency, I think it's
time to unify these. The callbacks add little value for our current
plan for how we would support a vkd3d-proton-native later.

I noticed that there were even callbacks for threads which is redundant code
at this point. Just use vkd3d_threads.h as-is.

A native OS handle is now wrapped inside a vkd3d_native_sync_handle.
It takes care of OS differences, and it's possible to signal a
native_sync_handle from d3d12_fence instead of HANDLE.

A HANDLE is wrapped into a vkd3d_native_sync_handle directly.
On Linux, a HANDLE is interpreted as an fd.
This supports both semaphore style and event-style.

Test code needed a way to create a HANDLE, so it was simplified to use
eventfd directly rather than a condition variable-based system.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>