vkd3d: Ensure that we don't have a floating presentation done event.
If we timeout for some reason, we risk entering a situation
where we are no longer guaranteed matching request and completion.

It could be solved with semaphores, but we definitely don't want to stay
out of sync more than one frame, so just block if we need to on next
present.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>