vkd3d: Implement extreme fast past for descriptor copies.
This introduces the idea of "embedded" mutable descriptors.

In this descriptor model, we can embed all information we need for
a given descriptor within the descriptor buffer payload itself.

This requires that the descriptor buffer implementation of the driver
supports some key properties:

- Size >= 32 (we need to reserve 5 LSBs to signal log2 offset to planar
  metadata in case that exists)
- Texel buffer and SSBO can be placed side by side in the mutable
  payload (these are usually 16 bytes each)
- Descriptor buffers can be bound at SSBO descriptor size alignment.
  We are clever here and bind the same descriptor buffer at an offset
  to achieve the packed texel buffer + SSBO descriptor.
- RTAS SRVs is implemented by viewing the descriptor buffer as an SSBO
  directly and shader applies a stride to load the BDA based on
  sizeof(mutable descriptor).
- UAV counters are implemented by repurposing the texel buffer.
  Could have used BDA, but it is somewhat less robust against app bugs
  when we alias texel buffer descriptor with it. We already had code
  paths in place to deal with the pre-BDA world, so this is fine.
- If Size is very large, like on RADV (64 byte) we can take
  advantage of that by embedding the metadata inside the mutable
  descriptor as well. This mostly just serves to save on CPU memory
  and avoids having to check for CPU -> CPU descriptor copies.

This allows us to massively optimize descriptor copies to the point
where the descriptor copy itself is an unrolled memcpy.

To further make this work well, we need to introduce vtable variants.
This improves performance quite a lot for descriptor copies, and is
overall cleaner.

In the descriptor copy micro-benchmark that copies one descriptor at a
time under ideal cache situations,
I observe 12.5 ns -> 2.5 ns per descriptor.

With the vtable variant system in place, it should be possible to add
special optimizations for implementations which do not hit this
optimization path.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>