idl: Get rid of INLINE_WRAPPER macros.
They break builds quite a lot since upstream widl and mingw64-tools widl
don't agree on how this is supposed to work.
If we need the macros, we have to macro away static which breaks other
things in the process.

The robust workaround is to roll our own wrappers for the functions
where it's needed.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>