vkd3d: Replace WSI fence wait with semaphore wait.
VkFence wait on Mesa is not a noop after all.
There can be very long stalls leading to very strange frame pacing
graphs when GPU is rendering slower than refresh rate in windowed modes.

Waiting for semaphores fixes this and with present_wait, there is little
reason to not use semaphores when there's a meaningful improvement in
jitter.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>