vkd3d: Get rid of VK_EXT_shader_viewport_index_layer.
Rely on Vulkan 1.2 features here.
Avoids problems with profiles layer since it will expose the EXT even if
we don't declare it, since it was promoted, but enabling the EXT
requires us to enable the features.

Spec says that enabling both features is equivalent to enabling the EXT,
so this should be okay.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>