vkd3d: Issue sfence on command list close as well.
There's a risk that if descriptors are copied in the thread that records
commands (common case), if submission happens in a completely different
thread, it's plausible that memory order does not hold.

This does not fix any known bug, so it's a hallucinated fix,
but might as well be defensive about it for more peace of mind.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>