vkd3d-ext: Replace several ID3D12GraphicsCommandListExt usages with d3d12_command_list_vkd3d_ext_iface
Replace several usages of the "incorrect" type in the
ID3D12GraphicsCommandListExt implementation. While ID3D12CommandListExt is the
literally correct type to use. The typedef
d3d12_command_list_vkd3d_ext_iface is needed, so that when an extension
of the ID3D12GraphicsCommandListExt interface is defined the pointer types will
be compatible.

Change the type used, as a follow-up change in this series will
introduce ID3D12GraphicsCommandListExt1.

Signed-off-by: Liam Middlebrook <lmiddlebrook@nvidia.com>