vkd3d: Apply UE5 workaround for mesh / bary support.
UE5 is broken and assumes that if mesh shaders are supported,
barycentrics are supported. It does not check for this, so PSOs will
start failing.

Work around this by disabling mesh shader support unless barycentrics
are also supported. RADV supports barycentrics now, so this is only a
real problem on older Mesa stable.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>