vkd3d: Move descriptor flags into metadata_view.
We really need to be able to squeeze all metadata for a UAV
into 16 bytes. This is necessary for upcoming change where we place SSBO
at 32 byte offset instead of 16 byte.

The use of goofy unions sharing a flags field is questionable,
but you gotta do what you gotta do ...

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>