vkd3d: Add new bindless flags for new RAW/TYPED split modes.
For the upcoming refactor, we'll have four possible pipeline layouts:

- Split mutable (ideal):
  First set: (texture, texel buffer)
  Second set: (UBO, SSBO)
  - May or may not have embedded mutable where typed and untyped live in
    separate halves of a descriptor.
- Full mutable (ANV until descriptor buffers are supported):
  - Cannot make this robust since we cannot emit sibling descriptors.
- Non-mutable:
  - *shrug*

The bindless info lookup is also accelerated here since for mutable we
statically know the info index. For non-mutable, fallback to a lookup.
We could in theory add static indices for that as well, but that gets
messy when we have to consider SSBO being used or not ...

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>