vkd3d: Refactor init command buffer system.
Sets us up for a world where we can have multiple sequences of init ->
app command buffers.
This will be important for hoisting DGC pre-patch work properly.

The main idea is that on indirect argument barriers, we can split the
command stream such that we can easily hoist work.
This is intended to remove some big bubbles in Starfield.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>