vkd3d: Add customizable scratch buffer pool count.
For indirect preprocess, RADV observes ~80 scratch buffers in New
Atlantis. Bump the maximum to 128 for DGC to avoid allocation churn.

Also fixes some additional churn on NV, but the block sizes there are
massive now, so it was not *as* bad, but it would still go a bit
beyond 32 blocks in New Atlantis. (Still pretty extreme, 1.5 GB just to
hold preprocess buffer ...)

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>