vkd3d-shader: Fully dup DXIL subobjects.
In deferred compilation, we cannot just have pointers straight into
the DXIL blob anymore. Owning the name pointer also allows us to rename it in a
collection import.

Add bool marking if the subobject is "owned" or not.
This is also a requirement for deferred compilation
since the COLLECTION will contain pre-parsed subobjects which can just
be referenced rather than duped.

Get rid of dxil_identifier as well. It was never used, and will be meaningless
in deferred compilation.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>