build: Replace stdcall-alias with kill-at, remove __declspec(dllexport)
This mirrors changes at DXVK [1], [2].

The intent in [1] was to prevent 32-bit export function names from
containing @Ordinals at the end.

This also turns out to be useful for compatibility with LLVM-based
toolchain, in particular when using LLD, because LLD does not recognize
--add-stdcall-alias but does recognize --kill-at.

As in [1], we use the system-provided d3d12 dependency instead of the
implib we generated, because MinGW LD insists on emitting implib with
ordinal suffixes (LLD does the correct thing here).

Another part of the change is to remove __declspec(dllexport). The .def
files we use specify the linkage already, so the attributes are obsolete.
This fixes LLD warnings like:

    ld.lld: warning: duplicate /export option: D3D12CreateDevice

There is another LLD compatibility issue here: while MSVC and MinGW will
give priority to the .def definitions, older versions of LLD would end up
honoring the attribute instead, resulting in default ordinals to be
exported instead of the specified one [3]. Not specifying dllexport in the
source fixes this.

[1]: https://github.com/doitsujin/dxvk/commit/a62117cd13618f98e5c0f0da27de9193cbc6dae2
[2]: https://github.com/doitsujin/dxvk/commit/904d3e6c9014bff7aa0843630b009c51879a1d06
[3]: https://github.com/llvm/llvm-project/issues/62329

Signed-off-by: Tatsuyuki Ishi <ishitatsuyuki@gmail.com>