d3d12: Work around duplicate LUIDs
Add a check when initializing a d3d12 physical device to also match the adapter device ID and vendor ID, since LUIDs don't seem to be unique on some hybrid systems.
Previously, the first available device was selected without checking for others matching the LUID, which would lead to the wrong device being selected on the aforementioned systems.