vkd3d: Mark all collection state objects created with DEPENDENCIES_ON_EXTERNAL_DEFINITIONS as deferred
This fixes a GPU hang with Warhammer 40,000: Darktide on NVIDIA GPUs. Currently this title creates
state object collections with the D3D12_STATE_OBJECT_FLAG_ALLOW_LOCAL_DEPENDENCIES_ON_EXTERNAL_DEFINITIONS
flag set that can successfully be compiled to a Vulkan RT pipeline. Later on the game creates another PSO
with additional state that needs to be included in the resulting Vulkan pipeline.

This doesn't work with the current vkd3d-proton implementation of DEPENDENCIES_ON_EXTERNAL_DEFINITIONS
because it expects all pipelines marked with this flag to fail the translation process into a Vulkan
pipeline. This change updates d3d12_state_object_init to always mark state object collections created
with DEPENDENCIES_ON_EXTERNAL_DEFINITIONS as deferred, even if it resulted in a valid pipeline object.