vkd3d: Add config flag to skip image heap clears.
These are typically used for resource heaps and don't need to be
cleared in most cases.
For RTV cases, they have to be initialized with a clear or discard,
and usually that will clear the memory anyway, so this flag is a softer
and safer variant of the full "skip clear" flag.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>