vkd3d: Implement frame rate limiter.
Follows the same logic as DXVK. Instead of delaying presentation, we instead delay
signaling of the frame latency event for hopefully consistent frame pacing on the
app side.

DXGI will call SetTargetFrameRate with a negative value when running in fullscreen
mode with a fixed refresh rate and Vsync enabled. The intent is to only engage the
frame rate limiter if the actual display refresh rate does not match the display
mode that the app is running at.

Signed-off-by: Philip Rebohle <philip.rebohle@tu-dortmund.de>